+++ src/io/io.hpp
@@ -64,7 +64,8 @@ auto h5_predtype() -> H5::PredType {
 }
 
 void write_output(const MeshState &mesh_state, GridStructure &mesh,
-                  const std::string &filename, int cycle, double time);
+                  ProblemIn *pin, const std::string &filename, int cycle,
+                  double time);
 
 void write_output(const MeshState &mesh_state, GridStructure &grid,
                   SlopeLimiter *SL, ProblemIn *pin, double time, int i_write);